Add git and vi to Dockerfile
While these are not strictly necessary for building and running
STELLOPT, they make interaction with the container environment much
easier.  Git in particular can also be used to capture version and
modification information at runtime.

For an editor, this change installs the "vim-tiny" package in order to
limit download sizes.