Add GSL-based optimizer routine
This routine uses GSL's trust-region solver to minimize chi^2.  It is
built along with STELLOPTV2/General and is linked into `xstelloptv2`.
The GSL library dependency has been added to the make rules for Docker
as well as to the Dockerfile.

The routine implements an interface `clmdif` which matches the `lmdif`
Fortran interface.