Remove any pre-existing make.inc
Previously, if the user had created a "make.inc" symlink in the host
copy of the repository, then the copy command in the Dockerfile would
overwrite the destination of that symlink in the container, making the
checkout unclean.  Now `--remove-destination` is used to ensure that
"make.inc" in the container is a simple copy of "SHARE/make_docker.inc".