Several files have commented lines added to them for readibility/testing.
The variable 'id_tag' has been added to stellopt_runtime. This variable is a copy of the original input filename, without modifications. This variable was added because 'id_string' is modified by stellopt, and SFINCS needs a copy of it original namelist, which is stored in the input file but does not get copied by stellopt when it creates new input files for each iteration.

stellopt_sfincs splits the MPI_COMM_MYWORLD communicator instead of MPI_COMM_WORLD